# cop-sys-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For COP-SYSTEM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/role": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get all roles",
                "operationId": "1ab8531c84d4e272efb9c5e930f755b1",
                "responses": {
                    "200": {
                        "description": "Roles list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RolesResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Add a new role",
                "operationId": "2a6380b210e6b77f5e9a9e5367c1200c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolesResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/role/{id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get a role by ID",
                "operationId": "829b48c5f588b576f3acbba0bb7df0c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the role to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolesResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Role"
                ],
                "summary": "Update an existing role",
                "operationId": "45c7c2c06095fd5349f231a5ee74cc6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the role to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolesResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Role"
                ],
                "summary": "Remove a role",
                "operationId": "5b73a204ab3eb7c8fbc7397e16f6bfcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the role to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role deleted successfully"
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            }
        },
        "/api/v1/ticket": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Get all tickets",
                "operationId": "019c9295ecb992bc766ecad43c27d4d0",
                "responses": {
                    "200": {
                        "description": "Ticket list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TicketsResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Add a new ticket",
                "operationId": "4813345492de867acbdd49437518da17",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTicketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketsResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/ticket/{id}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Get a ticket by ID",
                "operationId": "e4668e305907d267bff671354a35ec00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Update an existing ticket",
                "operationId": "3dd05588adf4df00aa638c15d0d75a4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Remove a ticket",
                "operationId": "618d2ba51f78ab25d6c0c690ff6963ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket deleted successfully"
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            }
        },
        "/api/v1/ticket/personnel/{personnel_id}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Get tickets assigned to a personnel",
                "operationId": "b8f7b2f9cabfa50888e1f03cd6b74157",
                "parameters": [
                    {
                        "name": "personnel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tickets retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TicketsResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personnel not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateRoleRequest": {
                "required": [
                    "role_title"
                ],
                "properties": {
                    "role_title": {
                        "type": "string"
                    },
                    "role_description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateRoleRequest": {
                "properties": {
                    "role_title": {
                        "type": "string"
                    },
                    "role_description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RolesResource": {
                "properties": {
                    "role_title": {
                        "type": "string"
                    },
                    "role_description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateTicketRequest": {
                "required": [
                    "description",
                    "title",
                    "price",
                    "assigned_personnel",
                    "vehicle",
                    "person"
                ],
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "assigned_personnel": {
                        "type": "string"
                    },
                    "vehicle": {
                        "type": "string"
                    },
                    "person": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateTicketRequest": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "assigned_personnel": {
                        "type": "string"
                    },
                    "vehicle": {
                        "type": "string"
                    },
                    "person": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TicketsResource": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "assigned_personnel": {
                        "type": "string"
                    },
                    "vehicle": {
                        "type": "string"
                    },
                    "person": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}